Added FindProgressWithY from Ellipse and Tests
Found problem with application of FindProgressWithX to change orbit.
To solve the problem, adding function FindProgressWithY to use when X value is too close to xAxis.
Also added tests to the new function.
Updated assert AreEqual to avoid tests fail when float are not exactly equals (0.41999998 != 0.4200017) to assert True with Mathf.Approximately