Show questions for quizz
Did a code simplification : old QcmVraiFaux and Manipulation are not addition and correction friendly, old version store the complete html code for EACH question in an array : take away memory for nothing, void is fast enough that pre-request is not necessary. Old versions did not store correction values nor took care of allowing or restricting interactions with solar system.

New version :
Bug for retrieval of float values, value given as param to the parser is apparently not correct. Currently looking for the cause of the error.