Fix Quizz for Vr
Would have preferred to use abstract class Detector so as not to have to duplicate almost all quiz scripts for the XR but currently don't have enough time left to do so. To optimize later.

Due to problem reaching host site for the start, retrieval of a question, created an emergency folder emulating the different html contents the application would have received.