Fix Bug : Interaction restrictions for quizz not activated
Resolved bug :
When the start quizz method is called, as well as the Next and StartCorrection methods, no new restriction is detected.

Reason :
Attribute "started" of quizz was never changed when the StartQuizz method was called.
Attribute "correctionDone" of quizzService was never changed after a new quastion or the start of a correction.